WIP core dependency workaround
